stack.yaml things are stuck with for now
What would be vastly more desirable is to avoid needing any files
outside the script file itself, including temporary ones. While
interpreters might do some package downloading or building or whatever,
they've already done the hard work of having different code for all of
the different systems things have to run on, which ideally would not
have to be reproduced here. Folding whatever package information is
needed into the script would be nice, but as best as I could tell, the
only way to accommodate the demand for specific version numbers to be
specified while also doing command-line evaluation was with the file on
the side. It's not hard to see where to intervene within stack's own
codebase for the command line option handling that provides a way to do
it for code in files but doesn't accept the same options for code in a
command-line argument, but carrying it out is enough work to not just
put out a branch of stack for that.